deepslide/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pymupdf>=1.23.8
camel-ai>=0.1.0
openai>=1.3.0
pydub>=0.25.1
SpeechRecognition>=3.10.0
jinja2>=3.1.2
httpx>=0.25.2
websockets>=12.0
Pillow>=10.0.0
colorama>=0.4.6
dashscope>=1.20.0
python-pptx>=1.0.0
playwright>=1.40.0
